refactor: PatchTeamModal 컴포넌트의 Props 의존성 제거
